# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THE GUIDE International Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentsPage"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/bulk-import:
    post:
      operationId: bulkImportStudents
      tags: [students]
      summary: Bulk import students from CSV/Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Certificates ──────────────────────────────────────────────
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      parameters:
        - name: approved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Add a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Contact ──────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string

    # ── Course ──────────────────────────────────────────────────
...
    Student:
      type: object
      required: [id, name, certificateNumber, courseName, completionDate, grade, status]
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        certificateNumber:
          type: string
        courseName:
          type: string
        completionDate:
          type: string
        grade:
          type: string
        status:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, certificateNumber, courseName, completionDate, grade]
      properties:
        name:
          type: string
        photoUrl:
          type: string
        certificateNumber:
          type: string
        courseName:
          type: string
        completionDate:
          type: string
        grade:
          type: string
        status:
          type: string
        email:
          type: string
        phone:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        photoUrl:
          type: string
        certificateNumber:
          type: string
        courseName:
          type: string
        completionDate:
          type: string
        grade:
          type: string
        status:
          type: string
        email:
          type: string
        phone:
          type: string
    StudentsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    BulkImportInput:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/StudentInput"
    BulkImportResult:
      type: object
      required: [imported, failed, total]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: string

    # ── Certificate ──────────────────────────────────────────────
...
    Testimonial:
      type: object
      required: [id, studentName, courseName, rating, review, approved]
      properties:
        id:
          type: integer
        studentName:
          type: string
        photoUrl:
          type: ["string", "null"]
        courseName:
          type: string
        rating:
          type: integer
        review:
          type: string
        designation:
          type: ["string", "null"]
        approved:
          type: boolean
        createdAt:
          type: string
    TestimonialInput:
      type: object
      required: [studentName, courseName, rating, review]
      properties:
        studentName:
          type: string
        photoUrl:
          type: string
        courseName:
          type: string
        rating:
          type: integer
        review:
          type: string
        designation:
          type: string
        approved:
          type: boolean
    TestimonialUpdate:
      type: object
      properties:
        studentName:
          type: string
        photoUrl:
          type: string
        courseName:
          type: string
        rating:
          type: integer
        review:
          type: string
        designation:
          type: string
        approved:
          type: boolean

    # ── Contact ──────────────────────────────────────────────────
...