# lib/api-zod/src/generated/types/bulkImportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THE GUIDE International Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentInput } from './studentInput';

export interface BulkImportInput {
  records: StudentInput[];
}