# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THE GUIDE International Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  certificateNumber: string;
  courseName: string;
  completionDate: string;
  grade: string;
  status: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt?: string;
}