# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THE GUIDE International Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  photoUrl?: string;
  certificateNumber: string;
  courseName: string;
  completionDate: string;
  grade: string;
  status?: string;
  email?: string;
  phone?: string;
}