# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THE GUIDE International Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  studentName: string;
  /** @nullable */
  photoUrl?: string | null;
  courseName: string;
  rating: number;
  review: string;
  /** @nullable */
  designation?: string | null;
  approved: boolean;
  createdAt?: string;
}